Add files via upload
I have added playlists. When you add a new playlist, you can select from your current existing voices to select a collection of them. When you use this Playlists as the current voice, it will randomly select a voice from that in order to give it a more conversational feel when reading comments or emails.